Implment front and back hierarchical model

+++ Bullet.cpp
@@ -6,8 +6,8 @@
 using namespace std;
 
 Bullet::Bullet(int _pos_x, int _pos_y, int _direction, int _length) {
-	pos_x = _pos_x;
-	pos_y = _pos_y;
+	pos_x = (float)_pos_x;
+	pos_y = (float)_pos_y;
 	direc = _direction;
 	move();
 	length = _length + 1;
